* reorganized mutexes * switched to spsc_queue as this structure does not have as many requirements on the content * split Transform into declaration and implementation * added type (static, dynamic) to Transform and removed passing this information over multiple methods as extra parameter * fixed segmentation fault caused by misalignment of Eigen structures in STL containers * fixed compiler warnings

+++ rsb/src/rct/impl/TransformCollectionConverter.cpp
@@ -22,7 +22,7 @@ using namespace rsb::converter;
 namespace rct {
 
 void TransformCollectionConverter::domainToRST(const vector<Transform>& transforms, FrameTransformCollection &ts) {
-    for(int i=0; i<transforms.size(); ++i) {
+    for(long unsigned int i=0; i<transforms.size(); ++i) {
         FrameTransform* t = ts.add_transforms();
         
 	transformConverter.domainToRST(transforms[i], *t);
